+++ ChangeLog.md
@@ -3,6 +3,7 @@ Unreleased
 
 * Windows builds now use `-D_SDL_main_h`. See https://github.com/haskell-game/sdl2/issues/139 for more discussion.
 * Support for event watching: `addEventWatch` and `delEventWatch`.
+* Several event payloads now have their `Window` fields modified to use `Maybe Window`, substituting `Nothing` for null pointers.
 * High-level structure for controller button state: `ControllerButtonState`.
 * High-level structure for controller buttons: `ControllerButton`.
 * High-level structure for controller connection: `ControllerDeviceConnection`.